pyproject.toml
[project]
name = "heimdall"
version = "0.2.0"
description = "Systems monitoring service"
authors = ["Aleksander S. <myth@overflow.no>"]
license = "MIT"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "aiosmtplib>=4.0.2",
    "aiosqlite>=0.21.0",
    "databases>=0.9.0",
    "fastapi>=0.116.1",
    "python-dateutil>=2.9.0.post0",
    "python-decouple>=3.8",
    "uvicorn>=0.35.0",
]

# Tools

[tool.ruff]
line-length = 120
target-version = 'py312'

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.12.12",
]
